refactor: generators
