+++ kewb/src/cube/mod.rs
@@ -1,6 +1,6 @@
 pub mod constants;
 pub mod cubie;
 pub mod facelet;
+pub mod generators;
 pub mod index;
 pub mod moves;
-pub mod utils;
